1.加入TescoStatys, 以Html呈现
2.runnow把schedule设为enable = true
3.修复Basket connector store/week 为空的时候不为null的bug

+++ PullingStatusTool/UserControl/UC_FileExpectMgr.cs
@@ -71,10 +71,15 @@ namespace PullingStatusTool.UserControl
 
         private void btn_Delete_Click(object sender, EventArgs e)
         {
-            DB_Helper db = new DB_Helper();
-            string id = gridView1.GetFocusedRowCellValue("c_expfileid").ToString().Trim();
-           if( db.deleteFileExpect(id))
-            getDS();
+             MessageBoxButtons messButton = MessageBoxButtons.OKCancel;
+            DialogResult dr = MessageBox.Show("Are you sure to delete?", "Confirm to delete", messButton);
+            if (dr == DialogResult.OK)
+            {
+                DB_Helper db = new DB_Helper();
+                string id = gridView1.GetFocusedRowCellValue("c_expfileid").ToString().Trim();
+                if (db.deleteFileExpect(id))
+                    getDS();
+            }
         }
         string id = "";
